Theme Sheet: Factor out ThemeFeaturesCard
