+++ client/my-sites/theme/main.jsx
@@ -8,7 +8,6 @@ import React from 'react';
 import { connect } from 'react-redux';
 import i18n from 'i18n-calypso';
 import titlecase from 'to-title-case';
-import { isArray } from 'lodash';
 import Gridicon from 'gridicons';
 
 /**
@@ -51,10 +50,10 @@ import PageViewTracker from 'lib/analytics/page-view-tracker';
 import DocumentHead from 'components/data/document-head';
 import { decodeEntities } from 'lib/formatting';
 import { getCanonicalTheme } from 'state/themes/selectors';
-import { isValidTerm } from 'my-sites/themes/theme-filters';
 import { recordTracksEvent } from 'state/analytics/actions';
 import { setThemePreviewOptions } from 'state/themes/actions';
 import ThemeNotFoundError from './theme-not-found-error';
+import ThemeFeaturesCard from './theme-features-card';
 
 const ThemeSheet = React.createClass( {
 	displayName: 'ThemeSheet',
@@ -294,14 +293,16 @@ const ThemeSheet = React.createClass( {
 	},
 
 	renderOverviewTab() {
-		const { isWpcomTheme, download } = this.props;
+		const { download, isWpcomTheme, siteSlug, taxonomies } = this.props;
 
 		return (
 			<div>
 				<Card className="theme__sheet-content">
 					{ this.renderDescription() }
 				</Card>
-				{ this.renderFeaturesCard() }
+				<ThemeFeaturesCard taxonomies={ taxonomies }
+					siteSlug={ siteSlug }
+					isWpcomTheme={ isWpcomTheme } />
 				{ download && <ThemeDownloadCard href={ download } /> }
 				{ isWpcomTheme && this.renderRelatedThemes() }
 				{ isWpcomTheme && this.renderNextTheme() }
@@ -432,36 +433,6 @@ const ThemeSheet = React.createClass( {
 		}
 
 		return renderedTab;
-	},
-
-	renderFeaturesCard() {
-		const { isWpcomTheme, siteSlug, taxonomies } = this.props;
-		if ( ! taxonomies || ! isArray( taxonomies.theme_feature ) ) {
-			return null;
-		}
-
-		const themeFeatures = taxonomies.theme_feature.map( function( item ) {
-			const term = isValidTerm( item.slug ) ? item.slug : `feature:${ item.slug }`;
-			return (
-				<li key={ 'theme-features-item-' + item.slug }>
-					{ ! isWpcomTheme
-						? <a>{ item.name }</a>
-						: <a href={ `/themes/filter/${ term }/${ siteSlug || '' }` }>{ item.name }</a>
-					}
-				</li>
-			);
-		} );
-
-		return (
-			<div>
-				<SectionHeader label={ i18n.translate( 'Features' ) } />
-				<Card>
-					<ul className="theme__sheet-features-list">
-						{ themeFeatures }
-					</ul>
-				</Card>
-			</div>
-		);
 	},
 
 	getDefaultOptionLabel() {
